+++ tests/functional/cr4me_link_test.php
@@ -49,7 +49,7 @@ class cr4me_link_posting_test extends \phpbb_functional_test_case
 				'Some more text with another cr - [url]http://kb.un1matr1x.de/kb.php?lang=da&show=3369&pw=[/url]');
 		$crawler = self::request('GET', "viewtopic.php?t={$post2['topic_id']}&sid={$this->sid}");
 
-		$this->assertEquals(2, $crawler->filter('.cr4me-image')->count());
+		$this->assertEquals(2, $crawler->filter('.cr4me-link-image')->count());
 
 		return $post;
 	}