+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Major 0
 
+## Minor 3
+
+### Revision 0
+
+* add coustom profilefields
+
 ## Minor 2
 
 ### Revision 1